feat(condo): DOMA-8523 updated button for open filters. Refactored button for reset filters
