+++ apps/condo/domains/acquiring/components/payments/PaymentsTable.tsx
@@ -1,4 +1,3 @@
-import { FilterFilled } from '@ant-design/icons'
 import { Col, Row, Space } from 'antd'
 import { Gutter } from 'antd/lib/grid/row'
 import dayjs, { Dayjs } from 'dayjs'
@@ -97,7 +96,6 @@ function usePaymentsSum (whereQuery) {
 const PaymentsTableContent: React.FC = (): JSX.Element => {
     const intl = useIntl()
     const SearchPlaceholder = intl.formatMessage({ id: 'filters.FullSearch' })
-    const FiltersButtonLabel = intl.formatMessage({ id: 'FiltersLabel' })
     const StartDateMessage = intl.formatMessage({ id: 'pages.condo.meter.StartDate' })
     const EndDateMessage = intl.formatMessage({ id: 'pages.condo.meter.EndDate' })
     const ConfirmTitle = intl.formatMessage({ id: 'component.TicketWarningModal.ConfirmTitle' })
@@ -143,7 +141,6 @@ const PaymentsTableContent: React.FC = (): JSX.Element => {
     const dateFilter = dateFilterValue ? dateFilterValue.map(el => el.toISOString()) : null
 
 
-
     const searchPaymentsQuery: Record<string, unknown> = {
         ...filtersToWhere({ advancedAt: dateFilter, ...filters }),
         organization: { id: organizationId },
@@ -178,7 +175,6 @@ const PaymentsTableContent: React.FC = (): JSX.Element => {
     }, [setDateRange])
 
 
-
     const onReset = useCallback(() => {
         setFiltersAreReset(true)
     }, [])
@@ -186,7 +182,7 @@ const PaymentsTableContent: React.FC = (): JSX.Element => {
     const {
         MultipleFiltersModal,
         ResetFiltersModalButton,
-        setIsMultipleFiltersModalVisible,
+        OpenFiltersButton,
     } = useMultipleFiltersModal(
         queryMetas,
         PaymentsFilterTemplate,
@@ -235,15 +231,7 @@ const PaymentsTableContent: React.FC = (): JSX.Element => {
                                         )
                                     }
                                     <Col>
-                                        <Button
-                                            secondary
-                                            type='sberBlack'
-                                            onClick={() => setIsMultipleFiltersModalVisible(true)}
-                                        >
-                                            <FilterFilled/>
-                                            {FiltersButtonLabel}
-                                            {appliedFiltersCount > 0 && ` (${appliedFiltersCount})`}
-                                        </Button>
+                                        <OpenFiltersButton />
                                     </Col>
                                 </Row>
                             </Col>
